requirements.txt
streamlit==1.35.0
python-dotenv==1.1.1
PyMuPDF==1.26.1
langchain==0.3.26
langchain-core==0.3.67
langchain-google-genai==2.0.10
langchain-pinecone==0.2.8
langchain-community==0.3.26
langchain-text-splitters==0.3.8
pinecone-client==6.0.0
google-generativeai==0.8.5
